move rsa gen into an internal package
